Fixed busy logic for SCA read/write commands
* Fixed busy logic for SCA read/write commands

Both read and write commands should wait for the busy flag to be cleared before being executed.
In particular, the current implementation of the Sea::read() function stored the `command` variable only once, and then directly tests the channel busy state with `barRead(sc_regs::SCA_RD_CMD.index)`. When the busy is cleared, the initial value of the `command` variable is passed to the `checkError()` function, and an exception is thrown because the busy bit was still set when the `barRead(sc_regs::SCA_RD_CMD.index)` was initially called to set the `command` variable.

The proposed change in the busy logic guarantees that the `command` variable gets fresher at each loop, until the busy flag is cleared or the timeout is reached.